refactor(mobile): comprehensive hooks architecture improvements
- Split monolithic hooks.ts (1288 lines) into domain-based modules
- Create mutation factory for announcement state changes (45% reduction)
- Scope cache invalidation to user context for all user-specific data
- Standardize useSession usage across screens
- Add global error handling via ToastContext
- Optimize calendar performance with memoization
- Fix FlashList ListHeader and renderItem memoization
