+++ mobile/src/hooks/index.ts
@@ -2,5 +2,6 @@ export { useDirectusAsset, getDirectusAssetUrl, buildDirectusAssetUrl } from './
 export { useNotifications } from './useNotifications';
 export { useUnreadSync } from './useUnreadSync';
 export { useSession, useSelectedChild } from './useSession';
+export { useQueryErrorHandler } from './useQueryErrorHandler';
 export type { SessionState, SessionPermissions } from './useSession';
 // ContentType is now exported from '../api/hooks' via useContentReadStatus
